Add files via upload
Wavegenerator draws a circle at the left and a wave at the right. By adjusting the amplitude A, the frequentie Frq and the phase Phs the circle and wave accordingly. All is working, exept the animation. I want the wave and circle to move by updating the phase Phs. I know I have to use the java Timer for that, but I don;t know how. The code is far from perfect, since I am a beginner. I hope someone can help me implemanting the java Timer in my clumsy code. Thank you.